moved clinicDao
