+++ mockingbird/src/main/kotlin/com/leftindust/mockingbird/graphql/types/GraphQLDoctor.kt
@@ -6,7 +6,7 @@ import com.expediagroup.graphql.generator.annotations.GraphQLName
 import com.leftindust.mockingbird.auth.GraphQLAuthContext
 import com.leftindust.mockingbird.dao.DoctorDao
 import com.leftindust.mockingbird.dao.EventDao
-import com.leftindust.mockingbird.dao.ReadClinicDao
+import com.leftindust.mockingbird.dao.clinic.ReadClinicDao
 import com.leftindust.mockingbird.dao.UserDao
 import com.leftindust.mockingbird.dao.entity.Doctor
 import com.leftindust.mockingbird.dao.patient.ReadPatientDao
